+++ src/zenml/zen_stores/migrations/versions/a39c4184c8ce_remove_secrets_manager_flavors.py
@@ -1,0 +1,31 @@
+"""remove secrets manager flavors [a39c4184c8ce].
+
+Revision ID: a39c4184c8ce
+Revises: 0.53.0
+Create Date: 2023-12-21 10:51:16.919710
+
+"""
+import sqlalchemy as sa
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "a39c4184c8ce"
+down_revision = "0.53.0"
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    """Upgrade database schema and/or data, creating a new revision."""
+    conn = op.get_bind()
+    meta = sa.MetaData(bind=op.get_bind())
+    meta.reflect(only=("flavor",))
+    flavors = sa.Table("flavor", meta)
+
+    # Remove all secrets manager flavors
+    conn.execute(flavors.delete().where(flavors.c.type == "secrets_manager"))
+
+
+def downgrade() -> None:
+    """Downgrade database schema and/or data back to the previous revision."""
+    pass